requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.2

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization and Logging
tensorboard>=2.13.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Utils
tqdm>=4.65.0
PyYAML>=6.0
scikit-learn>=1.3.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.1
pytest-cov>=4.1.0

# Optional: CUDA Support (uncomment if needed)
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# torch-scatter>=2.1.1
